Consolidate PatternRecognizer tests
